WEP-50 extract UserIdentity and tests
